plant-care-ai-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.4
langchain-google-genai==0.0.6
supabase==2.3.4
python-multipart==0.0.6
pillow==10.2.0
sentence-transformers==2.3.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.3
pydantic==2.5.3
httpx==0.26.0
